requirements.txt
asyncio==3.4.3
certifi==2025.1.31
numpy
pandas
scipy==1.15.2
setuptools==75.8.0
tzdata==2025.1
upstox-python-sdk
urllib3==2.3.0
websocket-client==1.8.0
websockets==15.0
six==1.17.0
requests==2.32.3
pandas_ta
TA-Lib
matplotlib
PyYAML
nats-py==2.14.0